Add relationships_info to feature model
