+++ featurebyte/models/feature_list.py
@@ -24,9 +24,8 @@ from featurebyte.models.base import (
     UniqueValuesConstraint,
     VersionIdentifier,
 )
-from featurebyte.models.feature import FeatureModel
+from featurebyte.models.feature import EntityRelationshipInfo, FeatureModel
 from featurebyte.models.feature_namespace import FeatureReadiness
-from featurebyte.models.relationship import RelationshipType
 from featurebyte.query_graph.graph import QueryGraph
 from featurebyte.query_graph.node import Node
 from featurebyte.query_graph.pruning_util import get_prune_graph_and_nodes
@@ -237,18 +236,6 @@ class FeatureCluster(FeatureByteBaseModel):
         List[Node]
         """
         return [self.graph.get_node_by_name(name) for name in self.node_names]
-
-
-class EntityRelationshipInfo(FeatureByteBaseModel):
-    """
-    Schema for entity relationship information (subset of existing RelationshipInfo)
-    """
-
-    id: PydanticObjectId = Field(default_factory=ObjectId, alias="_id", allow_mutation=False)
-    relationship_type: RelationshipType
-    entity_id: PydanticObjectId
-    related_entity_id: PydanticObjectId
-    relation_table_id: PydanticObjectId
 
 
 class FrozenFeatureListNamespaceModel(FeatureByteCatalogBaseDocumentModel):
